requirements.txt
-i https://pypi.org/simple
altair==4.2.2 ; python_version >= '3.7'
attrs==22.2.0 ; python_version >= '3.6'
blinker==1.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cachetools==5.3.0 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.0.1
click==8.1.3 ; python_version >= '3.7'
decorator==5.1.1 ; python_version >= '3.5'
entrypoints==0.4 ; python_version >= '3.6'
gitdb==4.0.10 ; python_version >= '3.7'
gitpython==3.1.31 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.0.0 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
joblib==1.2.0 ; python_version >= '3.7'
jsonschema==4.17.3 ; python_version >= '3.7'
llvmlite==0.39.1 ; python_version >= '3.7'
markdown-it-py==2.2.0 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
mdurl==0.1.2 ; python_version >= '3.7'
numba==0.56.4 ; python_version >= '3.7'
numpy==1.23.5
packaging==23.0 ; python_version >= '3.7'
pandas==1.5.3 ; python_version >= '3.8'
pillow==9.4.0 ; python_version >= '3.7'
plotly==5.13.1
protobuf==3.20.3 ; python_version >= '3.7'
pyarrow==11.0.0 ; python_version >= '3.7'
pydeck==0.8.0 ; python_version >= '3.7'
pygments==2.14.0 ; python_version >= '3.6'
pympler==1.0.1 ; python_version >= '3.6'
pynndescent==0.5.8
pyrsistent==0.19.3 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
rich==13.3.1 ; python_full_version >= '3.7.0'
scikit-learn==1.2.1 ; python_version >= '3.8'
scipy==1.10.1 ; python_version < '3.12' and python_version >= '3.8'
semver==2.13.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
setuptools==67.4.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.0 ; python_version >= '3.6'
streamlit==1.19.0
tenacity==8.2.2 ; python_version >= '3.6'
threadpoolctl==3.1.0 ; python_version >= '3.6'
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.12.0 ; python_version >= '3.5'
tornado==6.2 ; python_version >= '3.7'
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.5.0 ; python_version >= '3.7'
tzdata==2022.7 ; python_version >= '3.6'
tzlocal==4.2 ; python_version >= '3.6'
umap-learn==0.5.3
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
validators==0.20.0 ; python_version >= '3.4'
watchdog==2.3.1
zipp==3.15.0 ; python_version >= '3.7'
